+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Turret's Physics Patch
 
+## V 0.1.0
+
+Added impulse pulse
+
+changed the drag coefficient conversion
+    - factor from 0.01 to 0.02
+
 ## V 0.0.5
 
 initial release